Separate State and Condition code.
